+++ src/react/index.ts
@@ -2,5 +2,8 @@ export { useStorage } from './useStorage';
 export { useStorageState } from './useStorageState';
 export { useStorageReducer } from './useStorageReducer';
 export { useStorageSync } from './useStorageSync';
+export { useStorageEvent } from './useStorageEvent';
+export { StorageProvider, useStorageContext } from './StorageContext';
 
 export type { UseStorageStateResult } from './useStorageState';
+export type { StorageProviderProps } from './StorageContext';