+++ src/react/useStorageInstance.ts
@@ -1,0 +1,1 @@
+export { useStorageInstance } from './StorageContext';